Class-26 (Multimodal Chatbot-Small)(Omni edition)/requirements.txt
openai==0.28
pillow
PyMuPDF